inference-service/requirements.txt
gunicorn
pymongo
kafka-python
connexion[swagger-ui]